separate matcher-factory
