+++ src/device.h
@@ -25,27 +25,6 @@ namespace librealsense
     
 stream_interface * find_profile( rs2_stream stream, int index, std::vector< stream_interface * > const & profiles );
 
-class matcher_factory
-{
-public:
-    static std::shared_ptr< matcher > create( rs2_matchers matcher,
-                                                std::vector< stream_interface * > const & profiles );
-
-private:
-    static std::shared_ptr< matcher > create_DLR_C_matcher( std::vector< stream_interface * > const & profiles );
-    static std::shared_ptr< matcher > create_DLR_matcher( std::vector< stream_interface * > const & profiles );
-    static std::shared_ptr< matcher > create_DI_C_matcher( std::vector< stream_interface * > const & profiles );
-    static std::shared_ptr< matcher > create_DI_matcher( std::vector< stream_interface * > const & profiles );
-    static std::shared_ptr< matcher > create_DIC_matcher( std::vector< stream_interface* > const& profiles );
-    static std::shared_ptr< matcher > create_DIC_C_matcher( std::vector< stream_interface * > const & profiles );
-
-    static std::shared_ptr< matcher > create_identity_matcher( stream_interface * profiles );
-    static std::shared_ptr< matcher > create_frame_number_matcher( std::vector< stream_interface * > const & profiles );
-    static std::shared_ptr< matcher > create_timestamp_matcher( std::vector< stream_interface * > const & profiles );
-
-    static std::shared_ptr< matcher > create_timestamp_composite_matcher( std::vector< std::shared_ptr< matcher > > const & matchers );
-    static std::shared_ptr< matcher > create_frame_number_composite_matcher( std::vector< std::shared_ptr< matcher > > const & matchers );
-};
 
 class device : public virtual device_interface, public info_container
 {
